refactor(window): rename types to options
