+++ src/framework/theme/components/window/index.ts
@@ -1,4 +1,4 @@
 export * from './window.module';
 export * from './window.service';
 export * from './window-ref';
-export * from './window-types';
+export * from './window.options';
