+++ src/framework/theme/components/window/window.module.ts
@@ -6,7 +6,7 @@ import { NbCardModule } from '../card/card.module';
 import { NbWindowService } from './window.service';
 import { NbWindowsContainerComponent } from './windows-container.component';
 import { NbWindowComponent } from './window.component';
-import { NB_WINDOW_CONFIG, NbWindowConfig } from './window-types';
+import { NB_WINDOW_CONFIG, NbWindowConfig } from './window.options';
 import {
   NbWindowCloseIconComponent, NbWindowCollapseIconComponent, NbWindowExpandIconComponent, NbWindowMinimizeIconComponent,
 } from './window-icon.component';
