+++ src/framework/theme/components/window/window.options.ts
